requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
dictdiffer==0.8.1
docutils==0.16
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gspread==3.6.0
gspread-formatting==0.2.5
httplib2==0.18.1
idna==2.10
lockfile==0.12.2
oauth2client==4.1.3
oauthlib==3.1.0
pkg-resources==0.0.0
protobuf==3.12.4
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-daemon==2.2.4
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
secure-smtplib==0.1.1
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.10

lxml~=4.2.1
pip~=9.0.1
wheel~=0.30.0
ipython~=7.13.0
cryptography~=2.1.4
Pygments~=2.6.1
Pillow~=5.1.0
simplejson~=3.13.2
setuptools~=49.3.2
MarkupSafe~=1.0
keyring~=10.6.0
click~=7.1.2
numpy~=1.19.1
Flask~=1.1.2
SQLAlchemy~=1.3.19
pandas~=1.1.1